+++ lib/webdrivers/common.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require 'rubygems/package'
-require 'webdrivers/logger'
 require 'webdrivers/network'
 require 'webdrivers/system'
 require 'selenium-webdriver'
+require 'webdrivers/logger'
 require 'webdrivers/version'
 
 module Webdrivers
@@ -122,12 +122,12 @@ module Webdrivers
 
       private
 
+      def download_version
+        required_version == EMPTY_VERSION ? latest_version : required_version
+      end
+
       def download_url
-        @download_url ||= if required_version == EMPTY_VERSION
-                            downloads[downloads.keys.max]
-                          else
-                            downloads[normalize_version(required_version)]
-                          end
+        @download_url ||= direct_url(download_version).tap { |url| Webdrivers.logger.debug "#{file_name} URL: #{url}" }
       end
 
       def exists?
@@ -135,11 +135,7 @@ module Webdrivers
       end
 
       def correct_binary?
-        current_version == if required_version == EMPTY_VERSION
-                             latest_version
-                           else
-                             normalize_version(required_version)
-                           end
+        current_version == download_version
       rescue ConnectionError, VersionError
         driver_path if sufficient_binary?
       end